+++ svelte.config.js
@@ -1,9 +1,14 @@
+import preprocess from 'svelte-preprocess';
+
 /** @type {import('@sveltejs/kit').Config} */
 const config = {
 	kit: {
 		// hydrate the <div id="svelte"> element in src/app.html
 		target: '#svelte'
-	}
+	},
+	preprocess: preprocess({
+		postcss: true
+	}),
 };
 
 export default config;